+++ include/CenteredCamera.h
@@ -11,12 +11,12 @@ using namespace std;
 
 class CenteredCamera: public Camera {
 private:
-    double radius;
-    double pitch;
-    double yaw;
+    float radius;
+    float pitch;
+    float yaw;
 
 public:
-    CenteredCamera(double, double, double);
+    CenteredCamera(float, float, float);
     CenteredCamera();
     void lookAt();
     void eventUpdate(SDL_Event);
