+++ include/Constants.h
@@ -1,8 +1,8 @@
 #ifndef CONSTANTS_H
 #define CONSTANTS_H
 
-const double EPSILON = 0.0001;
-const double S_EPSILON = 0.00001;
-const double INF = 999999999999;
+const float EPSILON = 0.0001;
+const float S_EPSILON = 0.00001;
+const float INF = 999999999999;
 
 #endif
