+++ src/CenteredCamera.cpp
@@ -1,6 +1,6 @@
 #include "../include/CenteredCamera.h"
 
-CenteredCamera::CenteredCamera(double radius, double pitch, double yaw) : Camera() {
+CenteredCamera::CenteredCamera(float radius, float pitch, float yaw) : Camera() {
     this->radius = radius;
     this->pitch = pitch;
     this->yaw = yaw;
@@ -14,12 +14,12 @@ CenteredCamera::CenteredCamera() : Camera() {
 
 void CenteredCamera::lookAt() {
     Point eye = Point(1, 0, 0).rotate(Point(0,yaw,pitch)) * radius;
-    double centerX = 0;
-    double centerY = 0;
-    double centerZ = 0;
-    double upX = 0;
-    double upY = 1;
-    double upZ = 0;
+    float centerX = 0;
+    float centerY = 0;
+    float centerZ = 0;
+    float upX = 0;
+    float upY = 1;
+    float upZ = 0;
     gluLookAt(eye.getX(),eye.getY(),eye.getZ(),centerX,centerY,centerZ,upX,upY,upZ);
 }
